requirements.txt
Django==2.2.10
django-filter
djangorestframework
requests
